+++ examples/display/lcd/lcd_layer_blending/main/core/lvgl_ppa_blend.c
@@ -3,6 +3,7 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
+#include <inttypes.h>
 #include <string.h>
 #include "freertos/FreeRTOS.h"
 #include "freertos/semphr.h"
@@ -271,19 +272,18 @@ static void touchpad_read(lv_indev_t *indev_drv, lv_indev_data_t *data)
     esp_lcd_touch_handle_t tp = lv_indev_get_user_data(indev_drv);
     assert(tp);
 
-    uint16_t touchpad_x;
-    uint16_t touchpad_y;
     uint8_t touchpad_cnt = 0;
     /* Read data from touch controller into memory */
     esp_lcd_touch_read_data(tp);
 
     /* Read data from touch controller */
-    bool touchpad_pressed = esp_lcd_touch_get_coordinates(tp, &touchpad_x, &touchpad_y, NULL, &touchpad_cnt, 1);
-    if (touchpad_pressed && touchpad_cnt > 0) {
-        data->point.x = touchpad_x;
-        data->point.y = touchpad_y;
+    esp_lcd_touch_point_data_t touch_points[1] = {0};
+    esp_err_t ret = esp_lcd_touch_get_data(tp, touch_points, &touchpad_cnt, 1);
+    if (ret == ESP_OK && touchpad_cnt > 0) {
+        data->point.x = touch_points[0].x;
+        data->point.y = touch_points[0].y;
         data->state = LV_INDEV_STATE_PR;
-        ESP_LOGD(TAG, "Touch position: %d,%d", touchpad_x, touchpad_y);
+        ESP_LOGD(TAG, "Touch position: %" PRIu16 ",%" PRIu16, touch_points[0].x, touch_points[0].y);
     } else {
         data->state = LV_INDEV_STATE_REL;
     }