+++ components/display/lcd_touch/esp_lcd_touch_ili2118/CHANGELOG.md
@@ -1,5 +1,11 @@
 # ChangeLog
 
+## v0.1.1 - 2025-12-09
+
+### Changes:
+
+* Update docs and tests to use esp_lcd_touch_get_data instead of the deprecated esp_lcd_touch_get_coordinates
+
 ## v0.0.1 - 2025-01-07
 
 ### Enhancements: