+++ components/display/lcd_touch/esp_lcd_touch_st7123/CHANGELOG.md
@@ -1,5 +1,11 @@
 # ChangeLog
 
+## v1.0.1 - 2025-12-09
+
+### Changes:
+
+* Update docs and tests to use esp_lcd_touch_get_data instead of the deprecated esp_lcd_touch_get_coordinates
+
 ## v1.0.0 - 2024-08-12
 
 ### Enhancements: